feat(domain): Participant 추상 클래스를 만들어 Dealer와 Player가 상속 받도록 구현
